fix null coalesing

+++ httpdocs/src/build/js/classes/undo.js
@@ -8,6 +8,9 @@ class Undo {
       }
 
       if (event.key === 'Delete') {
+        if(cockpit.target.id === 'greentext' ) {
+          return;
+        }
         component.delete()
       }
     }.bind(this))
